restructure: HoC 파일 폴더 변경
