+++ src/entities/auth/ui/input/inputWithError.tsx
@@ -1,0 +1,6 @@
+import withError from '@/shared/components/input/HoC/withError';
+import Input from '@/shared/components/input';
+
+const InputWithMessage = withError(Input);
+
+export default InputWithMessage;
