--- src/shared/components/input/inputWithError.tsx
@@ -1,6 +1,0 @@
-import withError from '@/shared/lib/HOC/withError';
-import Input from '.';
-
-const InputWithMessage = withError(Input);
-
-export default InputWithMessage;
